# feat: Upgrade all dependencies and enable TypeScript declarations