--- a/package.json
+++ b/package.json
@@ -36,7 +36,7 @@
   },
   "license": "MIT",
   "dependencies": {
-    "chalk": "^5.3.0",
+    "chalk": "^5.6.2",
     "commander": "^11.1.0",
     "inquirer": "^9.2.12",
     "ora": "^7.0.1"
@@ -47,7 +47,7 @@
     "@vitest/coverage-v8": "^3.2.4",
     "@vitest/ui": "^3.2.4",
     "tsup": "^8.0.1",
-    "typescript": "^5.3.3",
+    "typescript": "^5.9.3",
     "vitest": "^3.2.4"
   },
   "engines": {