--- a/package.json
+++ b/package.json
@@ -37,7 +37,7 @@
   "license": "MIT",
   "dependencies": {
     "chalk": "^5.6.2",
-    "commander": "^11.1.0",
+    "commander": "^14.0.1",
     "inquirer": "^9.2.12",
     "ora": "^7.0.1"
   },