--- a/package.json
+++ b/package.json
@@ -39,7 +39,7 @@
     "chalk": "^5.6.2",
     "commander": "^14.0.1",
     "inquirer": "^9.2.12",
-    "ora": "^7.0.1"
+    "ora": "^9.0.0"
   },
   "devDependencies": {
     "@types/inquirer": "^9.0.7",