--- a/package.json
+++ b/package.json
@@ -38,11 +38,11 @@
   "dependencies": {
     "chalk": "^5.6.2",
     "commander": "^14.0.1",
-    "inquirer": "^9.2.12",
+    "inquirer": "^12.9.6",
     "ora": "^9.0.0"
   },
   "devDependencies": {
-    "@types/inquirer": "^9.0.7",
+    "@types/inquirer": "^9.0.9",
     "@types/node": "^24.6.1",
     "@vitest/coverage-v8": "^3.2.4",
     "@vitest/ui": "^3.2.4",