--- a/tsup.config.ts
+++ b/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'tsup'
 export default defineConfig({
   entry: ['src/index.ts'],
   format: ['esm'], // Use ESM for inquirer compatibility
-  dts: false, // Disable for now due to inquirer type issues
+  dts: true, // Enable TypeScript declarations
   splitting: false,
   sourcemap: true,
   clean: true,